+++ internal/api/handler/transaction/update/update.go
@@ -9,7 +9,7 @@ import (
 
 type RequestBody struct {
 	ID     int `json:"id" binding:"required"`
-	Amount float64 `json:"amount" binding:"required"`
+	Amount int64 `json:"amount" binding:"required"`
 }
 
 func (h *Handler) UpdateTransaction(c *gin.Context) {
