+++ internal/usecase/transaction/add/contract.go
@@ -3,5 +3,5 @@ package add
 import "context"
 
 type transactionRepo interface {
-	AddTransaction(ctx context.Context, categoryID int, amount float64) error
+	AddTransaction(ctx context.Context, categoryID int, amount int64) error
 }
